requirements.txt
# Telegram bot (async API)
python-telegram-bot>=20,<22

# OpenAI (new SDK)
openai>=1,<2

# Google Sheets
gspread>=6,<7
oauth2client>=4.1.3

# Environment variables
python-dotenv>=1.0.1
